review: fixnew23

+++ backend/controllers/cards.js
@@ -15,7 +15,9 @@ const deleteCard = async (req, res, next) => {
   try {
     const { id } = req.params;
     const card = await Card.findOne({ _id: id });
-
+    if (!card) {
+      throw new ErrorNotFound('Такой карточки нет');
+    }
     if (card.owner === req.user.id) {
       await Card.deleteOne(card);
       return res.status(200).send({ message: 'Карточка удалена' });
